refactor(Work): Column 제거, author를 User타입으로 변경.
- 사용자는 DefaultUser가 로그인한 상황을 가정함
- Work를 저장, 변경, 이동하는 Work의 상태가 변하는 것으로 해석하기 위해 Column을 제거
- 이에 따른 파급효과로 커밋의 덩어리가 커질 수 밖에 없었음

+++ BE/src/main/java/codesquad/todo/web/works/dto/WorkMoveResponse.java
@@ -5,10 +5,10 @@ import lombok.Getter;
 
 @Getter
 public class WorkMoveResponse {
-    private Integer order;
+    private Long order;
     private WorkStatus status;
 
-    public WorkMoveResponse(WorkStatus status, Integer order) {
+    public WorkMoveResponse(WorkStatus status, Long order) {
         this.status = status;
         this.order = order;
     }
